TST: misplaced arithmetic tests (#32275)
