+++ pandas/tests/frame/test_arithmetic.py
@@ -4,6 +4,7 @@ import operator
 
 import numpy as np
 import pytest
+import pytz
 
 import pandas as pd
 import pandas._testing as tm
@@ -771,3 +772,35 @@ def test_frame_single_columns_object_sum_axis_1():
     result = df.sum(axis=1)
     expected = pd.Series(["A", 1.2, 0])
     tm.assert_series_equal(result, expected)
+
+
+# -------------------------------------------------------------------
+# Unsorted
+#  These arithmetic tests were previously in other files, eventually
+#  should be parametrized and put into tests.arithmetic
+
+
+class TestFrameArithmeticUnsorted:
+    def test_frame_add_tz_mismatch_converts_to_utc(self):
+        rng = pd.date_range("1/1/2011", periods=10, freq="H", tz="US/Eastern")
+        df = pd.DataFrame(np.random.randn(len(rng)), index=rng, columns=["a"])
+
+        df_moscow = df.tz_convert("Europe/Moscow")
+        result = df + df_moscow
+        assert result.index.tz is pytz.utc
+
+        result = df_moscow + df
+        assert result.index.tz is pytz.utc
+
+    def test_align_frame(self):
+        rng = pd.period_range("1/1/2000", "1/1/2010", freq="A")
+        ts = pd.DataFrame(np.random.randn(len(rng), 3), index=rng)
+
+        result = ts + ts[::2]
+        expected = ts + ts
+        expected.values[1::2] = np.nan
+        tm.assert_frame_equal(result, expected)
+
+        half = ts[::2]
+        result = ts + half.take(np.random.permutation(len(half)))
+        tm.assert_frame_equal(result, expected)
