+++ pandas/tests/frame/test_period.py
@@ -4,10 +4,6 @@ from pandas import DataFrame, Index, PeriodIndex, period_range
 import pandas._testing as tm
 
 
-def _permute(obj):
-    return obj.take(np.random.permutation(len(obj)))
-
-
 class TestPeriodIndex:
     def test_as_frame_columns(self):
         rng = period_range("1/1/2000", periods=5)
@@ -42,15 +38,3 @@ class TestPeriodIndex:
 
         # it works!
         frame.to_string()
-
-    def test_align_frame(self):
-        rng = period_range("1/1/2000", "1/1/2010", freq="A")
-        ts = DataFrame(np.random.randn(len(rng), 3), index=rng)
-
-        result = ts + ts[::2]
-        expected = ts + ts
-        expected.values[1::2] = np.nan
-        tm.assert_frame_equal(result, expected)
-
-        result = ts + _permute(ts[::2])
-        tm.assert_frame_equal(result, expected)
